moving a couple more tests after upstream merge
